+++ R/internal.R
@@ -9,3 +9,11 @@
 isStrictPositiveInteger <- function(x){
   all(x > 0 & (floor(x) == x))
 }
+
+isSquareMatrix <- function(M){
+  is.matrix(M) && (nrow(M) == ncol(M))
+}
+
+isRealOrComplex <- function(M){
+  is.numeric(M) || is.complex(M)
+}