+++ src/RcppExports.cpp
@@ -6,6 +6,11 @@
 
 using namespace Rcpp;
 
+#ifdef RCPP_USE_GLOBAL_ROSTREAM
+Rcpp::Rostream<true>&  Rcpp::Rcout = Rcpp::Rcpp_cout_get();
+Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
+#endif
+
 // EigenR_det_real
 double EigenR_det_real(const Eigen::MatrixXd& M);
 RcppExport SEXP _EigenR_EigenR_det_real(SEXP MSEXP) {
@@ -345,6 +350,29 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// EigenR_exp_real
+Eigen::MatrixXd EigenR_exp_real(const Eigen::MatrixXd& M);
+RcppExport SEXP _EigenR_EigenR_exp_real(SEXP MSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type M(MSEXP);
+    rcpp_result_gen = Rcpp::wrap(EigenR_exp_real(M));
+    return rcpp_result_gen;
+END_RCPP
+}
+// EigenR_exp_cplx
+Rcpp::List EigenR_exp_cplx(const Eigen::MatrixXd& Re, const Eigen::MatrixXd& Im);
+RcppExport SEXP _EigenR_EigenR_exp_cplx(SEXP ReSEXP, SEXP ImSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type Re(ReSEXP);
+    Rcpp::traits::input_parameter< const Eigen::MatrixXd& >::type Im(ImSEXP);
+    rcpp_result_gen = Rcpp::wrap(EigenR_exp_cplx(Re, Im));
+    return rcpp_result_gen;
+END_RCPP
+}
 
 static const R_CallMethodDef CallEntries[] = {
     {"_EigenR_EigenR_det_real", (DL_FUNC) &_EigenR_EigenR_det_real, 1},
@@ -375,6 +403,8 @@ static const R_CallMethodDef CallEntries[] = {
     {"_EigenR_EigenR_UtDU_cplx", (DL_FUNC) &_EigenR_EigenR_UtDU_cplx, 2},
     {"_EigenR_EigenR_lsSolve_real", (DL_FUNC) &_EigenR_EigenR_lsSolve_real, 2},
     {"_EigenR_EigenR_lsSolve_cplx", (DL_FUNC) &_EigenR_EigenR_lsSolve_cplx, 4},
+    {"_EigenR_EigenR_exp_real", (DL_FUNC) &_EigenR_EigenR_exp_real, 1},
+    {"_EigenR_EigenR_exp_cplx", (DL_FUNC) &_EigenR_EigenR_exp_cplx, 2},
     {NULL, NULL, 0}
 };