+++ NEWS.md
@@ -1,3 +1,9 @@
+# EigenR 1.0.0.9000
+
+New matricial functions have been included: exponential, logarithm, cosine, 
+sine, hyperbolic cosine, hyperbolic sine, 
+
+
 # EigenR 1.0.0
 
 First release.